[logs] Fix TLS handshake hanging forever
Sometime when we try to open a new TLS connection the handshake hang forever
blocking the whole pipeline. It is a bug in golang that is known but not fixed:
* https://github.com/golang/go/issues/2263
* https://github.com/golang/go/issues/17708
The workaround is to manually add a timeout on the handshake like in
`tls.DialWithDialer`, see https://git.io/fjPZl.